Move ModelConfig from config/__init__.py to config/model.py
